Add services to upgrade structures and update resources

+++ Nexus/Nexus.Domain/Entities/Mine.cs
@@ -4,7 +4,7 @@
     {
         public short Id { get; set; }
         public required string Code { get; set; }
-        public decimal GainMultiplier { get; set; }
+        public int GainMultiplier { get; set; }
         public int ResourceId { get; set; }
 
         public virtual Resource Resource { get; set; }
